feature: add delete support
This commit adds support for deleting hosts. If a message is received
from the event queue, we will connect to the HSP database and delete
all records associated with the host.

This should run on its own pod since it will be listening on a
different queue than the egress queue. Some of the env vars will be
different as well.

Also, right now this is a very naïve approach that connects to the DB
for each delete event. After we get this deployed to CI and working,
I'll create a new PR to re-use the database connection.